"Changed 'deleteCard' method in CardService to 'deleteCardById' and updated corresponding call in CardController"
Explanation:
The diff shows that in the CardController, the method 'deleteCard' has been updated to 'deleteCardById' in the 'deleteList' method. Similarly, in the CardService, the 'deleteCard' method has been changed to 'deleteCardById' and the method signature has been updated to return a Promise of void instead of the previous return type. The implementation of the method remains the same, with the only change being the name of the method and the return type.